feat(dsg): extract graphql-decorator
